% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{print_table}
\alias{print_table}
\title{Print a table of data}
\usage{
print_table(dfo, SigF = NULL, caption = NULL, hline.after = c(0),
  secondRow = NULL, thirdRow = NULL, save = FALSE,
  fileout = "table_output.tex", ...)
}
\arguments{
\item{fileout}{}
}
\description{
Print a table of data
}
