% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_nekton}
\alias{table_nekton}
\title{Make formated table of Neuston nekton}
\usage{
table_nekton(df, caption = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Make formated table of Neuston nekton
}
