% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_neuston}
\alias{table_neuston}
\title{Make a formated neuston table}
\usage{
table_neuston(df, caption = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Make a formated neuston table
}
