% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_surfsamp}
\alias{table_surfsamp}
\title{Creates surface station summary table}
\usage{
table_surfsamp(df, caption = NULL, ...)
}
\arguments{
\item{filename}{Path of the file to be read in}

\item{saveLoc}{Folder to save the output}
}
\description{
Creates surface station summary table
}
